+++ src/Svea.WebPay.SDK.Tests/appsettings.Development.json
@@ -1,0 +1,14 @@
+﻿{
+  "Credentials": [
+    {
+      "MarketId": "SE",
+      "MerchantId": "201663",
+      "Secret": ""
+    },
+    {
+      "MarketId": "NO",
+      "MerchantId": "203362",
+      "Secret": ""
+    }
+  ]
+}
