+++ src/Svea.WebPay.SDK/CheckoutApi/Data.cs
@@ -1,11 +1,12 @@
 ﻿namespace Svea.WebPay.SDK.CheckoutApi
 {
+    using Svea.WebPay.SDK.CheckoutApi.Response;
     using System;
     using System.Text.Json.Serialization;
 
-    public class Data
+    public class Data : OrderData
     {
-        public Data() { }
+        public Data() : base() { }
 
         /// <summary>
         /// 
@@ -108,170 +109,18 @@
         public Data(MerchantSettings merchantSettings, Cart cart, Customer customer, Address shippingAddress, Address billingAddress, Gui gui, string locale, string currency,
             string countryCode, Presetvalue[] presetValues, string clientOrderNumber, long orderId, string emailAddress, string phoneNumber, PaymentType? paymentType,
             CheckoutOrderStatus status, object customerReference, bool? sveaWillBuyOrder, IdentityFlags identityFlags, object merchantData, PaymentInfo payment, string peppolId, GetOrderShippingInformation shippingInformation,
-            bool? recurring, string recurringToken = null)
+            bool? recurring, string recurringToken = null) : base(merchantSettings, cart, customer, shippingAddress, billingAddress, gui, locale, currency, countryCode,
+                presetValues, clientOrderNumber, orderId, emailAddress, phoneNumber, paymentType, status, customerReference,
+                sveaWillBuyOrder, identityFlags, merchantData, payment, peppolId, shippingInformation, recurring, recurringToken)
         {
-            MerchantSettings = merchantSettings;
-            Cart = cart;
-            Customer = customer;
-            ShippingAddress = shippingAddress;
-            BillingAddress = billingAddress;
+
             Gui = gui;
-            Locale = locale;
-            Currency = currency;
-            CountryCode = countryCode;
-            PresetValues = presetValues;
-            ClientOrderNumber = clientOrderNumber;
-            OrderId = orderId;
-            EmailAddress = emailAddress;
-            PhoneNumber = phoneNumber;
-            PaymentType = paymentType ?? CheckoutApi.PaymentType.Unknown;
-            Status = status;
-            CustomerReference = customerReference;
-            SveaWillBuyOrder = sveaWillBuyOrder;
-            IdentityFlags = identityFlags;
-            MerchantData = merchantData;
-            Payment = payment;
-            PeppolId = peppolId;
-            ShippingInformation = shippingInformation;
-            Recurring = recurring;
-            RecurringToken = recurringToken;
         }
 
-        /// <summary>
-        /// Specific merchant URIs
-        /// </summary>
-        /// <remarks>Required.</remarks>
-        [JsonInclude]
-        public MerchantSettings MerchantSettings { get; }
-
-        /// <summary>
-        /// Order rows.
-        /// </summary>
-        [JsonInclude]
-        public Cart Cart { get; }
-
-        /// <summary>
-        /// Identified customer of the order.
-        /// </summary>
-        [JsonInclude]
-        public Customer Customer { get; }
-
-        /// <summary>
-        /// Shipping address of identified customer.
-        /// </summary>
-        [JsonInclude]
-        public Address ShippingAddress { get; }
-
-        /// <summary>
-        /// Billing address of identified customer.
-        /// </summary>
-        [JsonInclude]
-        public Address BillingAddress { get; }
 
         [JsonInclude]
         public Gui Gui { get; }
 
-        /// <summary>
-        /// The current locale of the checkout, i.e.sv-SE etc.
-        /// </summary>
-        /// <remarks>Required.</remarks>
-        [JsonInclude]
-        public string Locale { get; }
 
-        /// <summary>
-        /// The current currency as defined by ISO 4217, i.e. SEK, NOK etc.
-        /// </summary>
-        /// <remarks>Required.</remarks>
-        [JsonInclude]
-        public string Currency { get; }
-
-        /// <summary>
-        /// Defined by two-letter ISO 3166-1 alpha-2, i.e. SE, DE, FI etc.
-        /// </summary>
-        [JsonInclude]
-        public string CountryCode { get; }
-
-        /// <summary>
-        /// A list containing the preset values, if any.
-        /// </summary>
-        [JsonInclude]
-        public Presetvalue[] PresetValues { get; }
-
-        [JsonInclude]
-        public string ClientOrderNumber { get; }
-
-        [JsonInclude]
-        public long OrderId { get; }
-
-        [JsonInclude]
-        public string EmailAddress { get; }
-
-        /// <summary>
-        /// The customer’s phone number
-        /// </summary>
-        [JsonInclude]
-        public string PhoneNumber { get; }
-
-        /// <summary>
-        /// The final payment method for the order. Will only have a value when the order is finalized, otherwise unknown.
-        /// </summary>
-        [JsonInclude]
-        public PaymentType? PaymentType { get; }
-
-        /// <summary>
-        /// The current state of the order
-        /// </summary>
-        [JsonInclude]
-        public CheckoutOrderStatus Status { get; }
-
-        /// <summary>
-        /// B2B Customer reference
-        /// </summary>
-        [JsonInclude]
-        public object CustomerReference { get; }
-
-        /// <summary>
-        /// True = Svea will buy this invoice. False = Svea will not buy this invoice. null = Selected payment method is not Invoice.
-        /// </summary>
-        [JsonInclude]
-        public bool? SveaWillBuyOrder { get; }
-
-        [JsonInclude]
-        public IdentityFlags IdentityFlags { get; }
-
-        /// <summary>
-        /// Metadata visible to the store
-        /// </summary>
-        /// <remarks>Optional. Cleaned up from Checkout database after 45 days.</remarks>
-        [JsonInclude]
-        public object MerchantData { get; }
-
-        /// <summary>
-        /// The final payment method for the order. Will only have a value when the order is finalized, otherwise null.
-        /// </summary>
-        [JsonInclude]
-        public PaymentInfo Payment { get; }
-
-        /// <summary>
-        /// A company’s ID in the PEPPOL network, which allows the company to receive PEPPOL invoices. A PEPPOL ID can be entered when placing a B2B order using the payment method invoice.
-        /// </summary>
-        [JsonInclude]
-        public string PeppolId { get; }
-
-        [JsonInclude]
-        public GetOrderShippingInformation ShippingInformation { get; }
-
-        /// <summary>
-        /// Indicates if the order is recurring order and will create a recurring token when order is finalized. Only applicable if merchant has recurring orders enabled.
-        /// </summary>
-        [JsonInclude]
-        public bool? Recurring { get; }
-
-
-        /// <summary>
-        /// Recurring token to be used for subsequent recurring orders. Only available when order is finalized. Only applicable if merchant has recurring orders enabled.
-        /// </summary>
-        [JsonInclude]
-        public string RecurringToken { get; }
     }
 }
